+++ src/css/stylehomeAndProfile.css
@@ -297,6 +297,7 @@ cursor: pointer;
   flex-direction: column;
   align-items: center;
   text-align: center;
+  height: fit-content;
 }
 .profile__infoUser{
   display: flex;
@@ -335,13 +336,13 @@ cursor: pointer;
 .CompanyDescription{
   font-style: italic;
   color:rgb(77, 75, 75);
+  overflow-wrap: break-word;
+  font-size: 0.9rem;
 }
 
 .companyProfileName{
   padding: 0.3em 0 0.6em 0;
 }
-
-
 
 
 /* responsive desktop */
@@ -364,6 +365,9 @@ cursor: pointer;
       padding-left: 1em;
   }
   
+  .profileB{
+    column-gap: 1em;
+  }
   .homeImgUser {
     width: 100%;
   }
@@ -383,5 +387,10 @@ cursor: pointer;
 
   .profile__cover{
   grid-column: 1/3;
+  }
+
+  #containerProfileListPosts{
+    grid-row: 2;
+    grid-column: 2/3;
   }
 }
