+++ src/firebase/fb-firestore.js
@@ -1,5 +1,4 @@
 // Save post in firestore
-
 const savePost = (username, userPost, date, userId, userPhoto, likes,url) =>
   firebase.firestore().collection("newPosts").add({ 
     username,
@@ -12,17 +11,11 @@ const savePost = (username, userPost, date, userId, userPhoto, likes,url) =>
     });
 
  // Save user in firestore
- 
  const saveUser = (user) =>
   firebase.firestore().collection("Users").add(user);
  
 
-  
 // Get Posts from Firestore to HTML on RealTime
- /* const onGetPosts = (callback) =>
-  firebase.firestore().collection("newPosts").orderBy('date', 'desc').onSnapshot(callback);*/
-
-
 const onGetPosts = (callback) =>
   firebase.firestore().collection("newPosts").orderBy('date', 'desc').onSnapshot((query) => {
     const data=[];
@@ -31,12 +24,11 @@ const onGetPosts = (callback) =>
         id:doc.id,
         ...doc.data()
       });
-      callback(data);
     })
+    callback(data);
   })
 
 // Get Users from firestore to HTML on Realtime
-
 const onGetUsers = (callback) =>
   firebase.firestore().collection("Users").orderBy('userName', 'asc').onSnapshot((query) => {
     const data=[];
@@ -45,24 +37,13 @@ const onGetUsers = (callback) =>
         id:doc.id,
         ...doc.data()
       });
-      callback(data);
     })
+    callback(data);
   })
-
- /* db.collection("cities").where("state", "==", "CA")
-  .onSnapshot((querySnapshot) => {
-      var cities = [];
-      querySnapshot.forEach((doc) => {
-          cities.push(doc.data().name);
-      });
-      console.log("Current cities in CA: ", cities.join(", "));
-  });*/
 
 
    const infoData = (id, callback) => firebase.firestore().collection('Users').where('userId', '==', id).onSnapshot(posts => {
-    console.log("prueba",posts.docs[0].id)
     return posts.docs.map(e => {
-        //console.log(e.data().userId);
       const dataUser= {
             ProfileId: posts.docs[0].id,
             userId : e.data().userId,
@@ -75,12 +56,29 @@ const onGetUsers = (callback) =>
             userLocation:e.data().userLocation,
             userDescription:e.data().userDescription,
       }
-       console.log(dataUser);
        callback(dataUser)
         })
   })
 
-
+  //  Get data from posts of active user
+  const getListPostUser = (id, callback) => firebase.firestore().collection('newPosts').where('userId', '==', id).onSnapshot(posts => {
+    const data=[];
+    posts.forEach((doc) => {
+      data.push({
+        id:doc.id,
+        ...doc.data()
+      });
+    })
+    callback(data.sort(function(a, b){
+      if (a.date < b.date) {
+        return 1;
+      }
+      if (a.date > b.date) {
+        return -1;
+      }
+      return 0;
+    }));
+  })
 
 
 // Delete published posts
@@ -97,4 +95,4 @@ const updatePost = (id, updatedPost) =>
 const updateUser = (id, updatedUser) =>
   firebase.firestore().collection("Users").doc(id).update(updatedUser);
 
-export { savePost, saveUser, onGetPosts, deletePosts, updatePost, getPost, onGetUsers, infoData, updateUser };
+export { savePost, saveUser, onGetPosts, deletePosts, updatePost, getPost, onGetUsers, infoData, updateUser, getListPostUser };
